--- src/components/pages/Map/Map.css
@@ -1,19 +1,0 @@
-.map-container {
-  margin-top: 75px;
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  width: 100%;
-}
-
-.geocoder {
-  position: absolute;
-  z-index: 1;
-  width: 50%;
-  left: 50%;
-  margin-left: -25%;
-  top: 10px;
-}
-.mapboxgl-ctrl-geocoder {
-  min-width: 100%;
-}
